Migrated to getNumberOfRounds
