+++ app/generators/contestantRoundScoreGenerator.tsx
@@ -1,5 +1,6 @@
 import { ITableRowData } from "../utils/wikiFetch"
 import { getTeamList, ITeam } from "../utils/wikiQuery"
+import { getNumberOfRounds } from '../utils/teamListUtils';
 import Team from '../models/Team'
 import LeagueStanding from '../models/LeagueStanding'
 
@@ -13,10 +14,7 @@ export async function generateContestantRoundScores(pageData: any, listOfContest
 
             return acc
         }, {})
-    const numberOfRounds = pageData.props.runners.reduce(
-        (acc: number, x: ITeam) => {
-            return x.eliminationOrder > acc ? x.eliminationOrder : acc
-        }, 0)
+    const numberOfRounds = getNumberOfRounds(pageData.props.runners)
 
     const result: LeagueStanding = new LeagueStanding()
